int -> double

+++ ft_raytracer.c
@@ -6,17 +6,17 @@
 /*   By: afaucher <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/02/11 19:41:11 by afaucher          #+#    #+#             */
-/*   Updated: 2014/03/27 10:53:39 by afaucher         ###   ########.fr       */
+/*   Updated: 2014/03/27 13:29:42 by afaucher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include			"raytracer.h"
 
 int					ft_getintervalue(t_obj *list, t_line *line,
-										t_obj **minobj, int min)
+										t_obj **minobj, double min)
 {
 	int				i;
-	int				value;
+	double			value;
 	t_vect			*dir;
 	t_point			*origin;
